feature(demos/auth-demo): adds auth demo
This commit adds an authentication demo app in the demo folder. The app
is a mock e-commerce app that requires the user to sign in to purchase
an item. The app demos a client-side Google sign in. It also contains a
straw-man custom sign-in that doesn't actually work to motivate the
Google sign in.